+++ CHANGELOG.md
@@ -27,6 +27,14 @@ If you like this project and find it useful, please consider giving it a star on
 
 You can also sponsor Tamer here https://buymeacoffee.com/6sjde6vkzl for his invaluable contribution to this project.
 
+## [2.2.6] - 2026-03-23
+
+### Fixed
+
+- [cover]: Fix command handler to Matterbridge v.3.7.0.
+
+<a href="https://www.buymeacoffee.com/luligugithub"><img src="https://matterbridge.io/assets/bmc-button.svg" alt="Buy me a coffee" width="80"></a>
+
 ## [2.2.5] - 2026-03-21
 
 ### Added